requirements.txt
academictorrents==2.3.3
addict==2.4.0
albumentations==1.4.11
apex==0.9.10dev
basicsr==1.4.2
cityscapesscripts==2.2.3
click==8.1.7
decord==0.6.0
diffusers==0.16.0
einops==0.8.0
fairscale==0.4.4
ftfy==6.2.0
horovod==0.28.1
imageio==2.34.2
iopath==0.1.10
ipython==8.12.3
kornia==0.6.12
legacy==0.1.7
lmdb==1.5.1
matplotlib==3.7.5
mc==0.0
more_itertools==10.3.0
numpy==1.24.1
omegaconf==2.3.0
onnx==1.16.1
opencv_python_headless==4.5.5.64
opendatasets==0.1.22
packaging==24.1
pandas==2.0.3
parrots==1.0.3
Pillow==10.4.0
prettytable==3.10.0
pudb==2024.1.1
pycocoevalcap==1.2
pycocotools==2.0.7
python_magic==0.4.27
pytorch_lightning==2.3.2
PyYAML==6.0.1
PyYAML==6.0.1
regex==2024.5.15
Requests==2.32.3
scikit_learn==1.3.2
scipy==1.14.0
skimage==0.0
spacy==3.2.6
tensorboardX==2.6.2.2
tensorflow==2.16.2
tifffile==2023.7.10
timm==0.4.12
torch==2.1.0+cu121
torch_tb_profiler==0.4.3
torchvision==0.16.0+cu121
tqdm==4.66.4
transformers==4.30.2
turbojpeg==0.0.2
typing_extensions==4.12.2
webdataset==0.2.86
yapf==0.40.2
